requirements.txt
pynetbox==6.1.3
python_powerdns==2.0.0
systemd-python==235
